+++ CHANGELOG.md
@@ -38,6 +38,7 @@ Changelog for NeoFS Node
 - Optimized GET/HEAD request forwarding (#3877)
 - Optimized netmap caching in node (#3966)
 - Store in metabase associated object ID in bytes instead of Base58 (#3971)
+- Optimized local RANGE request execution (#3967)
 
 ### Removed
 - `policer.max_workers` configuration (#3920)